move storefront page into pages folder
