+++ src/router.jsx
@@ -1,6 +1,6 @@
 import App from './App.jsx'
 import { RouterProvider, Route, createBrowserRouter, createRoutesFromElements } from 'react-router-dom'
-import Storefront from './Storefront.jsx'
+import Storefront from './pages/Storefront.jsx'
 import Login from './pages/Login.jsx'
 import ProfilePage from './pages/Profile.jsx'
 import AdminInventory from './AdminInventory.jsx'
